Add BoxAny::to_string()
