+++ lib/ffi/src/box_any.rs
@@ -1,7 +1,9 @@
 use super::{any::Any, arc::Arc};
+use crate::{call_function, cps, ByteString};
 use core::ops::Deref;
 
 #[repr(C)]
+#[derive(Clone)]
 pub struct BoxAny {
     inner: Arc<BoxAnyInner>,
 }
@@ -10,11 +12,28 @@ struct BoxAnyInner {
     value: Any,
 }
 
+extern "C" {
+    fn _pen_ffi_any_to_string(
+        stack: &mut cps::AsyncStack<ByteString>,
+        continuation: cps::ContinuationFunction<ByteString, ByteString>,
+        value: BoxAny,
+    ) -> cps::Result;
+}
+
 impl BoxAny {
     pub fn new(value: Any) -> Self {
         Self {
             inner: BoxAnyInner { value }.into(),
         }
+    }
+
+    pub async fn to_string(&self) -> ByteString {
+        call_function!(
+            fn(BoxAny) -> ByteString,
+            _pen_ffi_any_to_string,
+            self.clone()
+        )
+        .await
     }
 }
